+++ imx/examples/common/common.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"log"
-	"math/big"
 	"os"
 	"runtime"
 	"strings"
@@ -54,23 +53,16 @@ func CommonInitialise(configFilePath string) (context.Context, map[string]string
 }
 
 func NewStarkSigner(privateStarkKeyStr string) imx.L2Signer {
-	var starkPrivateKey *big.Int
 	var err error
-	if privateStarkKeyStr != "" {
-		var ok bool
-		starkPrivateKey, ok = new(big.Int).SetString(privateStarkKeyStr, 16)
-		if !ok {
-			log.Panicf("error in converting stark private key value from string to big.Int")
-		}
-	} else {
-		starkPrivateKey, err = stark.GenerateKey()
-		log.Println("Stark Private key: ", starkPrivateKey.String())
+	if privateStarkKeyStr == "" {
+		privateStarkKeyStr, err = stark.GenerateKey()
+		log.Println("Stark Private key: ", privateStarkKeyStr)
 		if err != nil {
 			log.Panicf("error in Generating Stark Private Key: %v\n", err)
 		}
 	}
 
-	l2signer, err := stark.NewSigner(starkPrivateKey)
+	l2signer, err := stark.NewSigner(privateStarkKeyStr)
 	if err != nil {
 		log.Panicf("error in creating StarkSigner: %v\n", err)
 	}